+++ tests/0200_analyzer/0210_api.py
@@ -1,38 +1,40 @@
 import pytest
 
-from dependency_comb.analyzer import DependenciesAnalyzer
+from dependency_comb.new_analyzer import DependenciesAnalyzer
 from dependency_comb.exceptions import AnalyzerAPIError
 
 from tests.utils import skip_api_condition
 
 
 # Skip marker decorator for tests depending on a API key usage
-api_allowed = pytest.mark.skipif(skip_api_condition, reason="API request is disabled")
+#api_allowed = pytest.mark.skipif(skip_api_condition, reason="API request is disabled")
+# TODO: Not sure yet if we still conditionnally disable tests with API call or not
+api_allowed = pytest.mark.skipif(False, reason="API request is disabled")
+
+
+#@api_allowed
+#def test_get_package_data_invalid_key():
+    #"""
+    #When given api key is invalid, the API respond with a Http 403 response
+    #"""
+    #analyzer = DependenciesAnalyzer(api_pause=None)
+
+    #with pytest.raises(AnalyzerAPIError) as excinfo:
+        #analyzer.get_package_data("project-composer")
+
+    #assert str(excinfo.value) == (
+        #"API responded a 403 error, your API key is probably invalid."
+    #)
+    #assert excinfo.value.http_status == 403
 
 
 @api_allowed
-def test_get_package_data_invalid_key(settings):
-    """
-    When given api key is invalid, the API respond with a Http 403 response
-    """
-    analyzer = DependenciesAnalyzer("dummy-key", api_pause=None)
-
-    with pytest.raises(AnalyzerAPIError) as excinfo:
-        analyzer.get_package_data("project-composer")
-
-    assert str(excinfo.value) == (
-        "API responded a 403 error, your API key is probably invalid."
-    )
-    assert excinfo.value.http_status == 403
-
-
-@api_allowed
-def test_get_package_data_invalid_name(settings):
+def test_get_package_data_invalid_name():
     """
     Request to API with a non existing package name should lead to an error from
     API
     """
-    analyzer = DependenciesAnalyzer(settings.api_key, api_pause=None)
+    analyzer = DependenciesAnalyzer(api_pause=None)
 
     with pytest.raises(AnalyzerAPIError) as excinfo:
         analyzer.get_package_data("combabcdenope")
@@ -45,15 +47,15 @@ def test_get_package_data_invalid_name(settings):
 
 
 @api_allowed
-def test_get_package_data_from_api(settings):
+def test_get_package_data_from_api():
     """
     Request to API for given package name should return proper JSON payload.
     """
-    analyzer = DependenciesAnalyzer(settings.api_key, api_pause=None)
+    analyzer = DependenciesAnalyzer(api_pause=None)
 
     # Currently the analyzer does not check for response status. It should do it
     # instead of coercing any response to json that lead to missunderstood issue
     payload = analyzer.get_package_data("project-composer")
 
-    assert payload["name"] == "project-composer"
-    assert payload["homepage"] == "https://github.com/sveetch/project-composer"
+    assert payload["info"]["name"] == "project-composer"
+    assert payload["info"]["home_page"] == "https://github.com/sveetch/project-composer"