+++ tests/0200_analyzer/0211_cache.py
@@ -1,18 +1,15 @@
-from dependency_comb.analyzer import DependenciesAnalyzer
+from dependency_comb.new_analyzer import DependenciesAnalyzer
 
 
 def test_get_package_data_from_cache(settings):
     """
     When cache file exists with the given package name it should be used without
     performing any request.
-
-    Here we use dummy key so any request should fails if cache system would not work
-    well.
     """
     cachedir = settings.fixtures_path / "api_cache"
-    analyzer = DependenciesAnalyzer("dummy-key", cachedir=cachedir)
+    analyzer = DependenciesAnalyzer(cachedir=cachedir)
 
     payload = analyzer.get_package_data("diskette")
 
-    assert payload["name"] == "diskette"
-    assert payload["latest_release_number"] == "0.3.5"
+    assert payload["info"]["name"] == "diskette"
+    assert payload["info"]["version"] == "0.3.6"